backend_api_service/requirements.txt
fastapi
uvicorn
python-dotenv
motor
requests
